+++ chainbase/src/main/java/org/tron/common/utils/DBConfig.java
@@ -117,4 +117,8 @@ public class DBConfig {
   @Getter
   @Setter
   private static long proposalExpireTime; // (ms)
+
+  @Getter
+  @Setter
+  private static String transactionHistoreSwitch;
 }
