+++ chainbase/src/main/java/org/tron/core/store/TransactionRetStore.java
@@ -1,4 +1,4 @@
-package org.tron.core.db;
+package org.tron.core.store;
 
 import com.google.protobuf.ByteString;
 import java.util.Objects;
@@ -8,9 +8,11 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
 import org.tron.common.utils.ByteArray;
+import org.tron.common.utils.DBConfig;
 import org.tron.core.capsule.TransactionInfoCapsule;
 import org.tron.core.capsule.TransactionRetCapsule;
-import org.tron.core.config.args.Args;
+import org.tron.core.db.TransactionStore;
+import org.tron.core.db.TronStoreWithRevoking;
 import org.tron.core.exception.BadItemException;
 import org.tron.protos.Protocol.TransactionInfo;
 
@@ -28,7 +30,7 @@ public class TransactionRetStore extends TronStoreWithRevoking<TransactionRetCap
 
   @Override
   public void put(byte[] key, TransactionRetCapsule item) {
-    if (BooleanUtils.toBoolean(Args.getInstance().getStorage().getTransactionHistoreSwitch())) {
+    if (BooleanUtils.toBoolean(DBConfig.getTransactionHistoreSwitch())) {
       super.put(key, item);
     }
   }
