+++ framework/src/main/java/org/tron/common/overlay/discover/node/NodeHandler.java
@@ -30,6 +30,7 @@ import org.tron.common.net.udp.message.discover.NeighborsMessage;
 import org.tron.common.net.udp.message.discover.PingMessage;
 import org.tron.common.net.udp.message.discover.PongMessage;
 import org.tron.common.overlay.discover.node.statistics.NodeStatistics;
+import org.tron.core.config.args.Args;
 
 @Slf4j(topic = "discover")
 public class NodeHandler {
@@ -189,7 +190,7 @@ public class NodeHandler {
       sendPong(msg.getTimestamp());
     }
     node.setP2pVersion(msg.getVersion());
-    if (!node.isConnectible()) {
+    if (!node.isConnectible(Args.getInstance().getNodeP2pVersion())) {
       changeState(State.NONACTIVE);
     } else if (state.equals(State.NONACTIVE) || state.equals(State.DEAD)) {
       changeState(State.DISCOVERED);
@@ -203,7 +204,7 @@ public class NodeHandler {
       getNodeStatistics().lastPongReplyTime.set(System.currentTimeMillis());
       node.setId(msg.getFrom().getId());
       node.setP2pVersion(msg.getVersion());
-      if (!node.isConnectible()) {
+      if (!node.isConnectible(Args.getInstance().getNodeP2pVersion())) {
         changeState(State.NONACTIVE);
       } else {
         changeState(State.ALIVE);
