+++ framework/src/main/java/org/tron/common/overlay/discover/node/NodeManager.java
@@ -118,7 +118,7 @@ public class NodeManager implements EventHandler {
   private void dbWrite() {
     List<Node> batch = new ArrayList<>();
     for (NodeHandler nodeHandler : nodeHandlerMap.values()) {
-      if (nodeHandler.getNode().isConnectible()) {
+      if (nodeHandler.getNode().isConnectible(Args.getInstance().getNodeP2pVersion())) {
         nodeHandler.getNode().setReputation(nodeHandler.getNodeStatistics().getReputation());
         batch.add(nodeHandler.getNode());
       }
@@ -165,7 +165,7 @@ public class NodeManager implements EventHandler {
   private void trimTable() {
     if (nodeHandlerMap.size() > NODES_TRIM_THRESHOLD) {
       nodeHandlerMap.values().forEach(handler -> {
-        if (!handler.getNode().isConnectible()) {
+        if (!handler.getNode().isConnectible(Args.getInstance().getNodeP2pVersion())) {
           nodeHandlerMap.remove(handler);
         }
       });
@@ -236,7 +236,7 @@ public class NodeManager implements EventHandler {
   public List<NodeHandler> getNodes(Predicate<NodeHandler> predicate, int limit) {
     List<NodeHandler> filtered = new ArrayList<>();
     for (NodeHandler handler : nodeHandlerMap.values()) {
-      if (handler.getNode().isConnectible() && predicate.test(handler)) {
+      if (handler.getNode().isConnectible(Args.getInstance().getNodeP2pVersion()) && predicate.test(handler)) {
         filtered.add(handler);
       }
     }
