+++ framework/src/main/java/org/tron/core/capsule/utils/TransactionUtil.java
@@ -16,11 +16,27 @@
 package org.tron.core.capsule.utils;
 
 import com.google.protobuf.ByteString;
+import java.util.ArrayList;
+import java.util.List;
+import java.util.Objects;
 import lombok.extern.slf4j.Slf4j;
+import org.apache.commons.lang3.StringUtils;
+import org.tron.common.runtime.InternalTransaction;
+import org.tron.common.runtime.ProgramResult;
+import org.tron.common.runtime.vm.LogInfo;
 import org.tron.common.utils.DecodeUtil;
+import org.tron.core.capsule.BlockCapsule;
+import org.tron.core.capsule.ReceiptCapsule;
 import org.tron.core.capsule.TransactionCapsule;
+import org.tron.core.capsule.TransactionInfoCapsule;
+import org.tron.core.config.args.Args;
+import org.tron.core.db.TransactionTrace;
+import org.tron.protos.Protocol;
 import org.tron.protos.Protocol.Transaction;
 import org.tron.protos.Protocol.Transaction.Contract;
+import org.tron.protos.Protocol.TransactionInfo;
+import org.tron.protos.Protocol.TransactionInfo.Log;
+import org.tron.protos.Protocol.TransactionInfo.code;
 import org.tron.protos.contract.BalanceContract.TransferContract;
 
 @Slf4j(topic = "capsule")
@@ -40,5 +56,88 @@ public class TransactionUtil {
 
     return new TransactionCapsule(transferContract,
         Contract.ContractType.TransferContract).getInstance();
+  }
+
+  public static TransactionInfoCapsule buildTransactionInfoInstance(TransactionCapsule trxCap, BlockCapsule block,
+      TransactionTrace trace) {
+
+    TransactionInfo.Builder builder = TransactionInfo.newBuilder();
+    ReceiptCapsule traceReceipt = trace.getReceipt();
+    builder.setResult(code.SUCESS);
+    if (StringUtils.isNoneEmpty(trace.getRuntimeError()) || Objects
+        .nonNull(trace.getRuntimeResult().getException())) {
+      builder.setResult(code.FAILED);
+      builder.setResMessage(ByteString.copyFromUtf8(trace.getRuntimeError()));
+    }
+    builder.setId(ByteString.copyFrom(trxCap.getTransactionId().getBytes()));
+    ProgramResult programResult = trace.getRuntimeResult();
+    long fee =
+        programResult.getRet().getFee() + traceReceipt.getEnergyFee()
+            + traceReceipt.getNetFee() + traceReceipt.getMultiSignFee();
+    ByteString contractResult = ByteString.copyFrom(programResult.getHReturn());
+    ByteString ContractAddress = ByteString.copyFrom(programResult.getContractAddress());
+
+    builder.setFee(fee);
+    builder.addContractResult(contractResult);
+    builder.setContractAddress(ContractAddress);
+    builder.setUnfreezeAmount(programResult.getRet().getUnfreezeAmount());
+    builder.setAssetIssueID(programResult.getRet().getAssetIssueID());
+    builder.setExchangeId(programResult.getRet().getExchangeId());
+    builder.setWithdrawAmount(programResult.getRet().getWithdrawAmount());
+    builder.setExchangeReceivedAmount(programResult.getRet().getExchangeReceivedAmount());
+    builder.setExchangeInjectAnotherAmount(programResult.getRet().getExchangeInjectAnotherAmount());
+    builder.setExchangeWithdrawAnotherAmount(
+        programResult.getRet().getExchangeWithdrawAnotherAmount());
+    builder.setShieldedTransactionFee(programResult.getRet().getShieldedTransactionFee());
+
+    List<Log> logList = new ArrayList<>();
+    programResult.getLogInfoList().forEach(
+        logInfo -> {
+          logList.add(LogInfo.buildLog(logInfo));
+        }
+    );
+    builder.addAllLog(logList);
+
+    if (Objects.nonNull(block)) {
+      builder.setBlockNumber(block.getInstance().getBlockHeader().getRawData().getNumber());
+      builder.setBlockTimeStamp(block.getInstance().getBlockHeader().getRawData().getTimestamp());
+    }
+
+    builder.setReceipt(traceReceipt.getReceipt());
+
+    if (Args.getInstance().isSaveInternalTx() && null != programResult.getInternalTransactions()) {
+      for (InternalTransaction internalTransaction : programResult
+          .getInternalTransactions()) {
+        Protocol.InternalTransaction.Builder internalTrxBuilder = Protocol.InternalTransaction
+            .newBuilder();
+        // set hash
+        internalTrxBuilder.setHash(ByteString.copyFrom(internalTransaction.getHash()));
+        // set caller
+        internalTrxBuilder.setCallerAddress(ByteString.copyFrom(internalTransaction.getSender()));
+        // set TransferTo
+        internalTrxBuilder
+            .setTransferToAddress(ByteString.copyFrom(internalTransaction.getTransferToAddress()));
+        //TODO: "for loop" below in future for multiple token case, we only have one for now.
+        Protocol.InternalTransaction.CallValueInfo.Builder callValueInfoBuilder =
+            Protocol.InternalTransaction.CallValueInfo.newBuilder();
+        // trx will not be set token name
+        callValueInfoBuilder.setCallValue(internalTransaction.getValue());
+        // Just one transferBuilder for now.
+        internalTrxBuilder.addCallValueInfo(callValueInfoBuilder);
+        internalTransaction.getTokenInfo().forEach((tokenId, amount) -> {
+          Protocol.InternalTransaction.CallValueInfo.Builder tokenInfoBuilder =
+              Protocol.InternalTransaction.CallValueInfo.newBuilder();
+          tokenInfoBuilder.setTokenId(tokenId);
+          tokenInfoBuilder.setCallValue(amount);
+          internalTrxBuilder.addCallValueInfo(tokenInfoBuilder);
+        });
+        // Token for loop end here
+        internalTrxBuilder.setNote(ByteString.copyFrom(internalTransaction.getNote().getBytes()));
+        internalTrxBuilder.setRejected(internalTransaction.isRejected());
+        builder.addInternalTransactions(internalTrxBuilder);
+      }
+    }
+
+    return new TransactionInfoCapsule(builder.build());
   }
 }
