+++ framework/src/test/java/org/tron/core/db/TransactionHistoryTest.java
@@ -13,6 +13,7 @@ import org.tron.core.capsule.TransactionInfoCapsule;
 import org.tron.core.config.DefaultConfig;
 import org.tron.core.config.args.Args;
 import org.tron.core.exception.BadItemException;
+import org.tron.core.store.TransactionHistoryStore;
 
 public class TransactionHistoryTest {
